services/web_services/hybrid_functions/requirements.txt
google-cloud-bigquery
pandas
python-dotenv
